Trait okaywal::LogManager

pub trait LogManager<M = StdFileManager>: Send + Sync + Debug + 'staticwhere
    M: FileManager,{
    // Required methods
    fn recover(&mut self, entry: &mut Entry<'_, M::File>) -> Result<()>;
    fn checkpoint_to(
        &mut self,
        last_checkpointed_id: EntryId,
        checkpointed_entries: &mut SegmentReader<M::File>,
        wal: &WriteAheadLog<M>
    ) -> Result<()>;

    // Provided method
    fn should_recover_segment(
        &mut self,
        _segment: &RecoveredSegment
    ) -> Result<Recovery> { ... }
}

Customizes recovery and checkpointing behavior for a WriteAheadLog.

Required Methods

fn recover(&mut self, entry: &mut Entry<'_, M::File>) -> Result<()>

Invoked once for each entry contained in all recovered segments within a WriteAheadLog.

Entry::read_chunk() can be used to read each chunk of data that was written via EntryWriter::write_chunk. The order of chunks is guaranteed to be the same as the order they were written in.

fn checkpoint_to( &mut self, last_checkpointed_id: EntryId, checkpointed_entries: &mut SegmentReader<M::File>, wal: &WriteAheadLog<M> ) -> Result<()>

Invoked each time the WriteAheadLog is ready to recycle and reuse segment files.

last_checkpointed_id is the id of the last entry that is being checkedpointed and removed from the log. If needed, checkpointed_entries can be used to iterate over all entries that are being checkpointed.

Shortly after this function returns, the entries stored within the file being checkpointed will no longer be accessible. To ensure ACID compliance of the underlying storage layer, all necessary changes must be fully synchronized to the underlying storage medium before this function returns.

Provided Methods

fn should_recover_segment( &mut self, _segment: &RecoveredSegment ) -> Result<Recovery>

When recovering a WriteAheadLog, this function is called for each segment as it is read. To allow the segment to have its data recovered, return Recovery::Recover. If you wish to abandon the data contained in the segment, return Recovery::Abandon.