Trait okaywal::LogManager

pub trait LogManager: Send + Sync + Debug + 'static {
    fn recover(&mut self, entry: &mut Entry<'_>) -> Result<()>;
    fn checkpoint_to(
        &mut self,
        last_checkpointed_id: EntryId,
        checkpointed_entries: &mut SegmentReader
    ) -> Result<()>;

    fn should_recover_segment(
        &mut self,
        _segment: &RecoveredSegment
    ) -> Result<Recovery> { ... }
}

Customizes recovery and checkpointing behavior for a WriteAheadLog.

Required Methods

fn recover(&mut self, entry: &mut Entry<'_>) -> Result<()>

Invoked once for each entry contained in all recovered segments within a WriteAheadLog.

Entry::read_chunk() can be used to read each chunk of data that was written via EntryWriter::write_chunk. The order of chunks is guaranteed to be the same as the order they were written in.

fn checkpoint_to(
&mut self,
last_checkpointed_id: EntryId,
checkpointed_entries: &mut SegmentReader
) -> Result<()>

Invoked each time the WriteAheadLog is ready to recycle and reuse segment files.

last_checkpointed_id is the id of the last entry that is being checkedpointed and removed from the log. If needed, checkpointed_entries can be used to iterate over all entries that are being checkpointed.

Provided Methods

fn should_recover_segment(
&mut self,
_segment: &RecoveredSegment
) -> Result<Recovery>

When recovering a WriteAheadLog, this function is called for each segment as it is read. To allow the segment to have its data recovered, return Recovery::Recover. If you wish to abandon the data contained in the segment, return Recovery::Abandon.